Commit-4

+++ OnlineCarSalesApp/src/main/java/com/cg/cars/repository/AddressRepository.java
@@ -1,15 +1,17 @@
 package com.cg.cars.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-
 import org.springframework.stereotype.Repository;
 import com.cg.cars.entities.Address;
 
 
 
+
 @Repository
-public interface AddressRepository extends JpaRepository<Address, Integer> {
+public interface AddressRepository extends JpaRepository<Address, Long> {
+	
+	
+
