+++ OnlineCarSalesApp/src/main/java/com/cg/cars/repository/CustomerRepository.java
@@ -5,10 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.cg.cars.entities.Customer;
 
-
-
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
